Classification/requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.55.3
fsspec==2024.2.0
grad-cam==1.5.4
huggingface-hub==0.27.0
idna==3.10
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
opencv-python==4.10.0.84
packaging==24.2
pillow==10.2.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
six==1.17.0
sympy==1.13.1
threadpoolctl==3.5.0
timm==1.0.12
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchvision==0.20.1+cu124
tqdm==4.67.1
ttach==0.0.3
typing_extensions==4.9.0
urllib3==2.3.0
